+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   to true, attempts to delete the cluster will fail. Set to false to disable
   delete protection.
 
+- New `cockroach_metric_export_prometheus_config` resource allows user to configure prometheus metric
+  export integration in AWS and GCP cloud providers.
+
+
 ## [1.5.0] - 2024-05-26
 
 - No changes.